requirements.txt
asgiref==3.4.1
certifi==2022.6.15
Django==3.2.5
mysqlclient==2.1.1
pip==22.1.2
psycopg2==2.8.6
PyMySQL==1.0.2
pytz==2022.1
setuptools==61.2.0
sqlparse==0.4.1
typing_extensions==4.3.0
wheel==0.37.1
djangorestframework~=3.14.0